backend/requirements.txt
# Framework API
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Alpaca Trading
alpaca-py==0.21.1

# LLM & AI
httpx>=0.24.0,<0.26
openai>=1.10.0

# Configuration
python-dotenv==1.0.1
pydantic-settings==2.2.1

# Database
supabase==2.4.0

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Scheduling & Async
apscheduler==3.10.4
aiohttp==3.9.3

# Testing
pytest>=7.0.0,<8
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
